requirements.txt
absl-py==0.11.0
astunparse==1.6.3
autocorrect==2.2.2
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.4
de-core-news-sm==2.3.0
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.8
lda==2.0.0
Markdown==3.3.3
matplotlib==3.3.3
murmurhash==1.0.4
nltk==3.5
np-utils==0.5.12.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
pbr==3.1.1
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyspellchecker==0.5.5
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.52.0
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.2.1
